Network manipulation library
In order to organize containers into networks, it is required to have
code duplicating some functionality of ifconfig(8) and route(8). These
utilities while present in the usual installations of FreeBSD might be
missing inside containers. Typical linux containers do not have these.

This change adds a library with following features:
* Interface structure with facilities to create, destroy, rename
  interfaces as well as add / del bridge members
* Functions to add / delete default routes.